# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MindWeaver AI",
        "title": "自动化自媒体工具-API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ctsch",
            "url": "http://www.example.com/support",
            "email": "xxxxxxx@xxx.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.0.106:14010",
    "basePath": "/api",
    "paths": {
...
        "/swaggers/doc": {
            "post": {
                "description": "Generates Markdown documentation for specified interfaces",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Swagger"
                ],
                "summary": "Generate API documentation from a Swagger document",
                "parameters": [
                    {
                        "description": "Request data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.GenerateDocRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.GenerateDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/swaggers/list": {
            "post": {
                "description": "Returns a list of all interfaces in the Swagger document",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Swagger"
                ],
                "summary": "List all interfaces in a Swagger document",
                "parameters": [
                    {
                        "description": "Request data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ListInterfacesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/utils.ApiInterface"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/tools/handle-llm-response": {
            "get": {
                "description": "解析html文件并判断是否出错",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "解析html文件并判断是否出错",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.ContextInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/tools/jsinspector": {
            "get": {
                "description": "解析html文件并判断是否出错",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tools"
                ],
                "summary": "解析html文件并判断是否出错",
                "parameters": [
                    {
                        "type": "string",
                        "description": "html文件的绝对路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.ContextInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.GenerateDocRequest": {
            "type": "object",
            "properties": {
                "api_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.ApiInterface"
                    }
                },
                "swagger_file": {
                    "type": "string"
                }
            }
        },
        "api.GenerateDocResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
        "api.ListInterfacesRequest": {
            "type": "object",
            "properties": {
                "swagger_file": {
                    "type": "string"
                }
            }
        },
...
        "base.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "services.ContextInfo": {
            "type": "object",
            "properties": {
                "current_file": {
                    "type": "string"
                },
                "cursor_position": {
                    "type": "integer"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "selected_code": {
                    "type": "string"
                }
            }
        },
...
        "utils.ApiInterface": {
            "type": "object",
            "properties": {
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}